Removed RepositoryProviderInterface and moved InMemoryRepositoryProvider to App\Service\Repository
