+++ config/services.yaml
@@ -74,10 +74,8 @@ services:
     App\Issues\StatusApi: '@App\Issues\GitHub\GitHubStatusApi'
     App\Issues\CommentsApiInterface: '@App\Issues\GitHub\GithubCommentsApi'
 
-    App\Repository\Provider\InMemoryRepositoryProvider:
+    App\Service\RepositoryProvider:
         arguments: [ '%repositories%' ]
-
-    App\Repository\Provider\RepositoryProviderInterface: '@App\Repository\Provider\InMemoryRepositoryProvider'
 
     subscriber.symfony_docs.milestone:
         class: App\Subscriber\MilestoneNewPRSubscriber
