+++ src/Issues/GitHubRequestHandler.php
@@ -4,7 +4,7 @@ namespace App\Issues;
 
 use App\Event\EventDispatcher;
 use App\Event\GitHubEvent;
-use App\Repository\Provider\RepositoryProviderInterface;
+use App\Service\RepositoryProvider;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\HttpFoundation\Request;
@@ -24,7 +24,7 @@ class GitHubRequestHandler implements LoggerAwareInterface
     private $repositoryProvider;
     private $container;
 
-    public function __construct(EventDispatcher $dispatcher, RepositoryProviderInterface $repositoryProvider)
+    public function __construct(EventDispatcher $dispatcher, RepositoryProvider $repositoryProvider)
     {
         $this->dispatcher = $dispatcher;
         $this->repositoryProvider = $repositoryProvider;
