+++ src/Issues/StatusApi.php
@@ -2,7 +2,7 @@
 
 namespace App\Issues;
 
-use App\Repository\Repository;
+use App\Model\Repository;
 
 /**
  * Sets and retrieves the status for an issue.
